pinctrl: msm: Remove explicit barriers from mmio ops
It turns out that every mmio operation is protected by a lock. So
replace every operation with a relaxed variant of it, and rename the
macro for it in case a new mmio operation does need a lock.

Replacing the operations with a relaxed variant reduces overhead for
pinctrl operations, which in turn benefits things like i2c.

Idea from Sultan Alsawaf <sultan@kerneltoast.com>

Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I7ab7ac086cadfb1bd240f5464daa2540b86dd821
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>