power: supply: qti: Fix real_type detection with A12 firmware
Xiaomi seems to have modified this enum in their driver as well
as in firmware (presumably adsp) of MIUI 13. Add a dummy element
to bump up the property id to what the firmware expects. Without
this change, real_type always reports as "Unknown", breaking
userspace batterysecret daemon.

Required for fast charging to work with MIUI 13 firmware.

Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I6633844f6ef677a7809e3c8eeef9c39d2a0b7f54
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>